commit source

+++ app/Observers/MemberTeamObserver.php
@@ -49,29 +49,30 @@ class MemberTeamObserver
     public function updated(MemberTeam $memberTeam)
     {
         Log::info(__CLASS__.' :: '.__FUNCTION__.' : '.'request_id='.$memberTeam->id);
-
-        $member = $memberTeam->member;
-        $team = $memberTeam->team;
-        $member->notify(new NewMember($team));
-
-        if($memberTeam->invited_by === null){
-            $admins = $memberTeam->team->admins;
-
-            foreach ($admins as $key => $value) {
-                $ntf = $value->admin->notifications()
+        if(!$memberTeam->isDirty('num_match')){
+            $member = $memberTeam->member;
+            $team = $memberTeam->team;
+            $member->notify(new NewMember($team));
+            if($memberTeam->invited_by === null){
+                $admins = $memberTeam->team->admins;
+    
+                foreach ($admins as $key => $value) {
+                    $ntf = $value->admin->notifications()
+                    ->where('data','LIKE','%"request_id":'.$memberTeam->id.'%')
+                    ->get();
+                    $ntf->map->delete();
+                }
+                $ntf = $memberTeam->team->captain->notifications()
+                ->where('data','LIKE','%"request_id":'.$memberTeam->id.'%')
+                ->get();
+                $ntf->map->delete();
+            }else{
+                $ntf = $memberTeam->member->notifications()
                 ->where('data','LIKE','%"request_id":'.$memberTeam->id.'%')
                 ->get();
                 $ntf->map->delete();
             }
-            $ntf = $memberTeam->team->captain->notifications()
-            ->where('data','LIKE','%"request_id":'.$memberTeam->id.'%')
-            ->get();
-            $ntf->map->delete();
-        }else{
-            $ntf = $memberTeam->member->notifications()
-            ->where('data','LIKE','%"request_id":'.$memberTeam->id.'%')
-            ->get();
-            $ntf->map->delete();
+
         }
     }
